refactor: update dependencies and improve code structure
- Removed outdated Babel dependencies and switched to ESLint with TypeScript support.
- Updated Rollup configuration to use modern plugins and removed unnecessary imports.
- Refactored string utility functions and moved them to a new file for better organization.
- Deleted TSLint configuration in favor of ESLint.
- Updated package.json to reflect new versions of dependencies and added new ones.
