+++ package.json
@@ -35,34 +35,29 @@
   "author": "Contentstack",
   "license": "MIT",
   "devDependencies": {
-    "@babel/preset-env": "^7.26.0",
     "@commitlint/cli": "^17.8.1",
     "@commitlint/config-conventional": "^17.8.1",
     "@types/jest": "^26.0.24",
-    "babel-core": "^6.26.3",
-    "babel-jest": "^29.7.0",
-    "babel-loader": "8.4.1",
-    "babel-preset-es2015": "^6.24.1",
     "commitizen": "^4.3.1",
-    "eslint": "^8.57.1",
+    "eslint": "^9.11.1",
+    "@typescript-eslint/eslint-plugin": "^8.6.0",
+    "@typescript-eslint/parser": "^8.6.0",
     "husky": "^8.0.3",
     "jest": "^29.7.0",
     "jest-coverage-badges": "^1.0.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-html-reporters": "^2.1.7",
     "jest-junit": "^15.0.0",
-    "jsdom": "^21.1.2",
+    "jsdom": "^25.0.0",
     "jsdom-global": "^3.0.2",
-    "prettier": "^2.8.8",
-    "rollup": "^2.79.2",
-    "rollup-plugin-json": "^4.0.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-sourcemaps": "^0.6.3",
+    "prettier": "^3.3.3",
+    "rimraf": "^6.0.1",
+    "rollup": "^4.21.3",
+    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-node-resolve": "^15.2.3",
     "rollup-plugin-typescript2": "^0.36.0",
     "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
-    "tslint": "^6.1.3",
-    "tslint-config-prettier": "^1.18.0",
     "typescript": "^4.9.5"
   },
   "config": {
@@ -74,10 +69,5 @@
     "hooks": {
       "pre-commit": "$TALISMAN_HOME/talisman_hook_script pre-commit && other-scripts"
     }
-  },
-  "babel": {
-    "presets": [
-      "es2015"
-    ]
   }
 }
