+++ rollup.config.js
@@ -1,11 +1,9 @@
-import resolve from 'rollup-plugin-node-resolve';
-import commonjs from 'rollup-plugin-sourcemaps';
-import sourceMaps from 'rollup-plugin-sourcemaps';
-import typescript from 'rollup-plugin-typescript2';
-import json from 'rollup-plugin-json';
+const resolve = require('@rollup/plugin-node-resolve').default;
+const typescript = require('rollup-plugin-typescript2');
+const json = require('@rollup/plugin-json').default;
 
-import pkg from './package.json'
-export default {
+const pkg = require('./package.json');
+module.exports = {
     input: 'src/index.ts',
     output: [
         {
@@ -26,16 +24,11 @@ export default {
             useTsconfigDeclarationDir: true,
             objectHashIgnoreUnknownHack: false      
         }),
-        // Allow bundling cjs modules (unlike webpack, rollup doesn't understand cjs)
-        commonjs(),
 
         // Allow node_modules resolution, so you can use 'external' to control
         // which external modules to include in the bundle
         // https://github.com/rollup/rollup-plugin-node-resolve#usage
         resolve(),
-
-        // Resolve source maps to the original source
-        sourceMaps(),
     ],
-}
+};
