+++ src/helper/string-utils.ts
@@ -1,0 +1,9 @@
+export function replaceHtmlEntities(text: string): string {
+  return text
+    .replace(/&/g, '&amp;')
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;');
+}
+
+export const forbiddenAttrChars = ['"', "'", '>','<', '/', '='];
