--- tslint.json
@@ -1,6 +1,0 @@
-{
-    "extends": ["tslint:recommended", "tslint-config-prettier"],
-    "rules": {
-        "no-console": false
-    }
- }
